Optimize go usage in Dockerfile
Not everything we build needs go, and we really don't need to invalidate
caches every time we change go versions.
So instead of using `FROM golang`, just pull in debain:buster and mount
in `go` where it is needed.

For the dev target (used for `make shell`, tests, etc) we'll go ahead
and copy go into the image.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>