api: clarify "v" option behavior in the container remove endpoint
Current description of the "v" option doesn't explain what happens to
the volumes that are still in use by other containers. Turns out that
the only volumes that are removed are unnamed ones[1].

Perhaps a good way of clarifying this behavior would be adapting the
description from "docker rm --help".

As for the docs/api/v1.*.yaml changes — they seem to be applicable,
since the origin of this behavior dates way back to the 2016 or v1.11[2].

[1]: https://github.com/moby/moby/blob/a24a71c50f34d53710cccaa4d5e5f62169c5e1dc/daemon/mounts.go#L34-L38
[2]: https://github.com/moby/moby/commit/dd7d1c8a02d8693aa4f381f82c5bbdcad9a5ff58

Signed-off-by: Nikolay Edigaryev <edigaryev@gmail.com>
(cherry picked from commit 7afef854ed14bba2cc07d53e6f21455056053a35)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>