ipamutils: generate default pools on first use, instead of init()
This package defines the `NetworkToSplit` type. Curently importing
the package also initializes the default address pools (which may
not be needed, e.g., when used in the client.

This patch changes initialization to happen on first use, instead
of `init()`, to allow the package to be imported without initializing
these pools.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>