WIP: Fix permissions for newly created dirs in ADD
edb62a3ace8c4303822a391b38231e577f8c2ee8 made it so permissions are set
for existing dirs, it also inadvertantly made it so `Chmod` is called
against (previously) non-existing dirs.
This had a side-effect that dir permissions are different than in
previous version of Docker when using `ADD` to add a tarball during
`docker build` (There may be other cases of this).

This makes it so we don't set perms for paths that we created as part of
`MkdirAll` since the perms should already be the ones we expect.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>