Remove VALIDATE_REPO
This variable was used to allow running CI on the docker/engine
repository (and our internal "docker-ee" repositories).

Now that release branches have moved back to the upstream repository,
there's no longer a need to set these overrides.

I kept the variable in `hack/validate/.validate`, in case someone
wants to override it manually, but it's unused otherwise.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>