Add Samuel Karp as a maintainer
Samuel has been a long-time supporter of the container ecosystem.  He is an active security advisor for the containerd project, has spent a lot of time with the awslogs logging driver, and has been helping with review and triage of pull requests here for a while, and I believe his experience with both Go and containers at large would be a very valuable addition to the maintainer team.

Signed-off-by: Tianon Gravi <admwiggin@gmail.com>