staticcheck: SA1019: strings.Title is deprecated (in Go 1.18)
Replace it with golang.org/x/text/cases.Title()

    daemon/logger/templates/templates.go:23:14: SA1019: strings.Title is deprecated: The rule Title uses for word boundaries does not handle Unicode punctuation properly. Use golang.org/x/text/cases instead. (staticcheck)
    "title":    strings.Title,
    ^
    pkg/plugins/pluginrpc-gen/template.go:67:9: SA1019: strings.Title is deprecated: The rule Title uses for word boundaries does not handle Unicode punctuation properly. Use golang.org/x/text/cases instead. (staticcheck)
    return strings.Title(s)
    ^

TBH, I'm not sure if we really need this; most strings we generate will be ASCII,
so this would only be in case it's used for some user-provided string. We also
don't have a language to use, so would be using the "default".

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>