daemon/config: DefaultShmSize: minor tweak and improve docs
I had to check what the actual size was, so added it to the const's documentation.

While at it, also made use of it in a test, so that we're testing against the expected
value, and changed one alias to be consistent with other places where we alias this
import.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 821b4d410857debdfb539db18952f86feb7609f8)
Signed-off-by: Bjorn Neergaard <bjorn.neergaard@docker.com>